+++ src/clj/matthiasn/systems_toolbox/sente.clj
@@ -2,9 +2,9 @@
   (:gen-class)
   (:require
     [clojure.tools.logging :as log]
-    [ring.middleware.defaults]
+    [ring.middleware.defaults :as rmd]
     [ring.util.response :refer [resource-response response content-type]]
-    [compojure.core :refer (defroutes GET POST)]
+    [compojure.core :refer (routes GET POST)]
     [compojure.route :as route]
     [matthiasn.systems-toolbox.component :as comp]
     [clojure.core.async :refer [<! chan put! mult tap pub sub timeout go-loop sliding-buffer]]
@@ -13,7 +13,7 @@
     [taoensso.sente.server-adapters.immutant :refer (sente-web-server-adapter)]
     [taoensso.sente.packers.transit :as sente-transit]))
 
-(def ring-defaults-config (assoc-in ring.middleware.defaults/site-defaults [:security :anti-forgery]
+(def ring-defaults-config (assoc-in rmd/site-defaults [:security :anti-forgery]
                                     {:read-token (fn [req] (-> req :params :csrf-token))}))
 
 (defn user-id-fn
@@ -33,21 +33,22 @@
 (def host (get (System/getenv) "HOST" "localhost"))
 (def port (get (System/getenv) "PORT" "8888"))
 
-(defn mk-state
+(defn sente-comp-fn
   "Return clean initial component state atom."
-  [index-page-fn]
+  [{:keys [index-page-fn middleware]} ]
   (fn [put-fn]
     (let [ws (sente/make-channel-socket! sente-web-server-adapter {:user-id-fn user-id-fn
                                                                    :packer (sente-transit/get-flexi-packer :edn)})
-          {:keys [ch-recv ajax-get-or-ws-handshake-fn ajax-post-fn]} ws]
-      (defroutes my-routes
-                 (GET "/" [] (content-type (response (index-page-fn false)) "text/html"))
-                 (GET "/chsk" req (ajax-get-or-ws-handshake-fn req))
-                 (POST "/chsk" req (ajax-post-fn req))
-                 (route/resources "/")
-                 (route/not-found "Page not found"))
-      (let [my-ring-handler (ring.middleware.defaults/wrap-defaults my-routes ring-defaults-config)
-            server (immutant/run my-ring-handler :host host :port port)]
+          {:keys [ch-recv ajax-get-or-ws-handshake-fn ajax-post-fn]} ws
+          cmp-routes (routes
+                       (GET "/" req (content-type (response (index-page-fn req)) "text/html"))
+                       (GET "/chsk" req (ajax-get-or-ws-handshake-fn req))
+                       (POST "/chsk" req (ajax-post-fn req))
+                       (route/resources "/")
+                       (route/not-found "Page not found"))]
+      (let [ring-handler (rmd/wrap-defaults cmp-routes ring-defaults-config)
+            wrapped-in-middleware (if middleware (middleware ring-handler) ring-handler)
+            server (immutant/run wrapped-in-middleware :host host :port port)]
         (log/info "Immutant-web is listening on port" port "on interface" host))
       (sente/start-chsk-router! ch-recv (make-handler ws put-fn))
       ws)))
@@ -68,12 +69,15 @@
 (defn cmp-map
   "Creates server-side WebSockets communication component map."
   {:added "0.3.1"}
-  [cmp-id index-page-fn]
-  {:cmp-id           cmp-id
-   :state-fn         (mk-state index-page-fn)
-   :all-msgs-handler all-msgs-handler
-   :opts             {:watch :connected-uids
-                      :snapshots-on-firehose false}})
+  [cmp-id cfg-map-or-index-page-fn]
+  (let [cfg-map (if (map? cfg-map-or-index-page-fn)
+                  cfg-map-or-index-page-fn
+                  {:index-page-fn cfg-map-or-index-page-fn})]
+    {:cmp-id           cmp-id
+     :state-fn         (sente-comp-fn cfg-map)
+     :all-msgs-handler all-msgs-handler
+     :opts             {:watch :connected-uids
+                        :snapshots-on-firehose false}}))
 
 (defn component
   "Creates server-side WebSockets communication component map. Unlike in other components, this function is still